Fixes #23: Pass electron version to checkers
If the directory and ASAR loaders find a package.json file, they now
check if there is a dependency on `electron` and if so, remember its
version number.
That version number is passed to the checkers' `match()` functions.

I modified the `electron.asar` in the tests folder to include a
minimal package.json file, containing:

```
{
  "devDependencies": {
    "electron": "6.1.11"
  }
}
```

The `test_loader.js` test now also tests if this version is read
correctly.